+++ tests/test_dataset.py
@@ -2,12 +2,13 @@ import os
 import signal
 import pytest
 from sys import modules
-from unittest.mock import patch, Mock, PropertyMock
+from unittest.mock import patch, Mock, PropertyMock, call
 from egcg_core import constants as c
 from integration_tests.mocked_data import MockedSamples, MockedRunProcess
 from tests.test_analysisdriver import TestAnalysisDriver, NamedMock
 from analysis_driver.exceptions import AnalysisDriverError, SequencingRunError
 from analysis_driver.dataset import Dataset, RunDataset, SampleDataset, MostRecentProc
+from analysis_driver.notification import LimsNotification
 from analysis_driver.tool_versioning import Toolset
 
 ppath = 'analysis_driver.dataset.'
@@ -71,7 +72,8 @@ class TestDataset(TestAnalysisDriver):
     @patched_initialise
     @patch(ppath + 'MostRecentProc.start')
     @patch(ppath + 'MostRecentProc.retrieve_entity')
-    def test_start(self, mocked_retrieve_entity, mocked_start, mocked_update):
+    @patch(ppath + 'SampleDataset.lims_ntf', new_callable=PropertyMock)
+    def test_start(self, mocked_lims_ntf, mocked_retrieve_entity, mocked_start, mocked_update):
         self.dataset.most_recent_proc.entity['status'] = c.DATASET_PROCESSING
         with patched_get_run, patched_is_ready:
             with pytest.raises(AssertionError):
@@ -83,7 +85,8 @@ class TestDataset(TestAnalysisDriver):
 
     @patched_finish
     @patch(ppath + 'MostRecentProc.retrieve_entity')
-    def test_succeed(self, mocked_retrieve_entity, mocked_finish):
+    @patch(ppath + 'SampleDataset.lims_ntf', new_callable=PropertyMock)
+    def test_succeed(self, mocked_lims_ntf, mocked_retrieve_entity, mocked_finish):
         with patched_get_run, patched_is_ready:
             with pytest.raises(AssertionError):
                 self.dataset.succeed()
@@ -94,7 +97,8 @@ class TestDataset(TestAnalysisDriver):
 
     @patched_finish
     @patch(ppath + 'MostRecentProc.retrieve_entity')
-    def test_fail(self, mocked_retrieve_entity, mocked_finish):
+    @patch(ppath + 'SampleDataset.lims_ntf', new_callable=PropertyMock)
+    def test_fail(self, mocked_lims_ntf, mocked_retrieve_entity, mocked_finish):
         with patched_get_run, patched_is_ready:
             with pytest.raises(AssertionError):
                 self.dataset.succeed()
@@ -106,7 +110,8 @@ class TestDataset(TestAnalysisDriver):
 
     @patched_finish
     @patch(ppath + 'MostRecentProc.retrieve_entity')
-    def test_abort(self, mocked_retrieve_entity, mocked_finish):
+    @patch(ppath + 'SampleDataset.lims_ntf', new_callable=PropertyMock)
+    def test_abort(self, mocked_lims_ntf, mocked_retrieve_entity, mocked_finish):
         self.dataset.abort()
         mocked_finish.assert_called_with(c.DATASET_ABORTED)
 
@@ -401,6 +406,33 @@ class TestSampleDataset(TestDataset):
             {'run_id': 'another_run_id', 'clean_q30_bases_r1': 110, 'clean_q30_bases_r2': 135}
         ]
         self.dataset._data_threshold = 1000000000
+
+    @patched_initialise
+    @patch(ppath + 'MostRecentProc.start')
+    @patch(ppath + 'MostRecentProc.retrieve_entity')
+    @patch(ppath + 'SampleDataset.lims_ntf', new_callable=PropertyMock)
+    def test_start(self, mocked_lims_ntf, mocked_retrieve_entity, mocked_start, mocked_update):
+        with patched_get_run:
+            self.dataset.start()
+            mocked_lims_ntf.assert_has_calls([call(), call().start_sample_pipeline()])
+
+    @patched_finish
+    @patch(ppath + 'MostRecentProc.retrieve_entity')
+    @patch(ppath + 'SampleDataset.lims_ntf', new_callable=PropertyMock)
+    def test_succeed(self, mocked_lims_ntf, mocked_retrieve_entity, mocked_finish):
+        with patched_get_run:
+            self.dataset.most_recent_proc.entity['status'] = c.DATASET_PROCESSING
+            self.dataset.succeed()
+            mocked_lims_ntf.assert_has_calls([call(), call().assign_next_and_advance_step()])
+
+    @patched_finish
+    @patch(ppath + 'MostRecentProc.retrieve_entity')
+    @patch(ppath + 'SampleDataset.lims_ntf', new_callable=PropertyMock)
+    def test_fail(self, mocked_lims_ntf, mocked_retrieve_entity, mocked_finish):
+        self.dataset.most_recent_proc.entity['status'] = c.DATASET_PROCESSING
+        self.dataset.fail(1)
+        self.assertTrue(mocked_lims_ntf.called)
+        mocked_lims_ntf.assert_has_calls([call(), call().remove_sample_from_workflow()])
 
 
 class TestMostRecentProc(TestAnalysisDriver):